pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "commitments-reconciler"
version = "0.1.0"
description = "Integration tests and utilities for commitments reconciliation"
authors = [{name = "AutoGenerated", email = "dev@example.com"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
dependencies = [
    "openpyxl>=3.0",
    "python-docx>=0.8",
]

[project.optional-dependencies]
test = ["pytest"]

[tool.setuptools]
packages = ["commitments_reconciler"]
